+++ src/pipe/gui/handler/ArcHandler.java
@@ -58,9 +58,10 @@ public class ArcHandler extends PetriNetObjectHandler {
 				if (e.isControlDown()) {
 					CreateGui.getCurrentTab().getUndoManager().addNewEdit(
 							arc.getArcPath().insertPoint(
-									new Point2D.Float(
+									new Point2D.Double(
 									    arc.getX() + e.getX(),
-                                        arc.getY() + e.getY()),
+                                        arc.getY() + e.getY()
+                                    ),
 									e.isAltDown()
                             )
                     );
